src/test/java/practice/leetcode/kataArrayDiff/SomeTest.java
package practice.leetcode.kataArrayDiff;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class SomeTest {
    @Test
    public void sampleTests() {
        assertArrayEquals(new int[]{2}, Some.arrayDiff(new int[]{1, 2}, new int[]{1}));
        assertArrayEquals(new int[]{2, 2}, Some.arrayDiff(new int[]{1, 2, 2}, new int[]{1}));
        assertArrayEquals(new int[]{1}, Some.arrayDiff(new int[]{1, 2, 2}, new int[]{2}));
        assertArrayEquals(new int[]{1, 2, 2}, Some.arrayDiff(new int[]{1, 2, 2}, new int[]{}));
        assertArrayEquals(new int[]{}, Some.arrayDiff(new int[]{}, new int[]{1, 2}));
    }
}